+++ programs/locked-voter/src/instructions/cast_vote.rs
@@ -74,6 +74,7 @@ impl<'info> Validate<'info> for CastVote<'info> {
             self.proposal.get_state()? == ProposalState::Active,
             "proposal must be active"
         );
+        invariant!(self.escrow.is_not_in_dispute(), "Escrow is disputed");
         Ok(())
     }
 }