+++ programs/locked-voter/src/instructions/dispute_instructions/mod.rs
@@ -1,0 +1,8 @@
+pub mod open_dispute;
+pub use open_dispute::*;
+pub mod resolve_dispute;
+pub use resolve_dispute::*;
+pub mod withdraw_fund_from_dispute;
+pub use withdraw_fund_from_dispute::*;
+pub mod close_dispute;
+pub use close_dispute::*;