+++ programs/locked-voter/src/instructions/partial_unstaking_instructions/mod.rs
@@ -1,0 +1,6 @@
+pub mod merge_partial_unstaking;
+pub mod open_partial_unstaking;
+pub mod withdraw_partial_unstaking;
+pub use merge_partial_unstaking::*;
+pub use open_partial_unstaking::*;
+pub use withdraw_partial_unstaking::*;