+++ tests/govern/claimReward.ts
@@ -316,7 +316,7 @@ describe("Locked voter", () => {
         for (const keypair of userKeypairs) {
             const wallet = new Wallet(keypair);
             const voterProgram = createLockedVoterProgram(wallet, LOCKED_VOTER_PROGRAM_ID);
-            const [escrow, _bump] = deriveEscrow(locker, wallet.publicKey, LOCKED_VOTER_PROGRAM_ID);
+            const escrow = deriveEscrow(locker, wallet.publicKey, LOCKED_VOTER_PROGRAM_ID);
 
             await voterProgram.methods
                 .newEscrow()
@@ -344,6 +344,15 @@ describe("Locked voter", () => {
             );
 
             await voterProgram.methods
+                .extendLockDuration(maxStakeDuration)
+                .accounts({
+                    escrow,
+                    escrowOwner: wallet.publicKey,
+                    locker,
+                })
+                .rpc();
+
+            await voterProgram.methods
                 .increaseLockedAmount(lockAmount)
                 .accounts({
                     escrow,
@@ -352,15 +361,6 @@ describe("Locked voter", () => {
                     payer: voterProgram.provider.publicKey,
                     sourceTokens: rewardATA,
                     tokenProgram: TOKEN_PROGRAM_ID,
-                })
-                .rpc();
-
-            await voterProgram.methods
-                .extendLockDuration(maxStakeDuration)
-                .accounts({
-                    escrow,
-                    escrowOwner: wallet.publicKey,
-                    locker,
                 })
                 .rpc();
         }
@@ -469,7 +469,7 @@ describe("Locked voter", () => {
             const wallet = new Wallet(keypair);
             const voterProgram = createLockedVoterProgram(wallet, LOCKED_VOTER_PROGRAM_ID);
             const governProgram = createGovernProgram(wallet, GOVERN_PROGRAM_ID);
-            const [escrow, _eBump] = deriveEscrow(locker, wallet.publicKey, LOCKED_VOTER_PROGRAM_ID);
+            const escrow = deriveEscrow(locker, wallet.publicKey, LOCKED_VOTER_PROGRAM_ID);
             const vote = await getOrCreateVote(proposal, governProgram);
             await voterProgram.methods
                 .castVote(VoteSide.For)