+++ tests/govern/claimRewardOptionProposal.ts
@@ -326,7 +326,7 @@ describe("Locked voter", () => {
         for (const keypair of userKeypairs) {
             const wallet = new Wallet(keypair);
             const voterProgram = createLockedVoterProgram(wallet, LOCKED_VOTER_PROGRAM_ID);
-            const [escrow, _bump] = deriveEscrow(locker, wallet.publicKey, LOCKED_VOTER_PROGRAM_ID);
+            const escrow = deriveEscrow(locker, wallet.publicKey, LOCKED_VOTER_PROGRAM_ID);
 
             await voterProgram.methods
                 .newEscrow()
@@ -354,6 +354,14 @@ describe("Locked voter", () => {
             );
 
             await voterProgram.methods
+                .extendLockDuration(maxStakeDuration)
+                .accounts({
+                    escrow,
+                    escrowOwner: wallet.publicKey,
+                    locker,
+                })
+                .rpc();
+            await voterProgram.methods
                 .increaseLockedAmount(lockAmount)
                 .accounts({
                     escrow,
@@ -362,15 +370,6 @@ describe("Locked voter", () => {
                     payer: voterProgram.provider.publicKey,
                     sourceTokens: rewardATA,
                     tokenProgram: TOKEN_PROGRAM_ID,
-                })
-                .rpc();
-
-            await voterProgram.methods
-                .extendLockDuration(maxStakeDuration)
-                .accounts({
-                    escrow,
-                    escrowOwner: wallet.publicKey,
-                    locker,
                 })
                 .rpc();
         }
@@ -479,7 +478,7 @@ describe("Locked voter", () => {
             const wallet = new Wallet(keypair);
             const voterProgram = createLockedVoterProgram(wallet, LOCKED_VOTER_PROGRAM_ID);
             const governProgram = createGovernProgram(wallet, GOVERN_PROGRAM_ID);
-            const [escrow, _eBump] = deriveEscrow(locker, wallet.publicKey, LOCKED_VOTER_PROGRAM_ID);
+            const escrow = deriveEscrow(locker, wallet.publicKey, LOCKED_VOTER_PROGRAM_ID);
             const vote = await getOrCreateVote(proposal, governProgram);
             await voterProgram.methods
                 .castVote(VoteSide.For)