+++ tests/utils/index.ts
@@ -3,3 +3,4 @@ export * from "./merkle-tree";
 export * from "./program";
 export * from "./helper";
 export * from "./types";
+export * from "./token";